+++ src/Areas/LoadTesting/Models/LoadTest/InputArtifacts.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace AzureMcp.Areas.LoadTesting.Models.LoadTest;
+
+public class InputArtifacts
+{
+    [JsonPropertyName("testScriptFileInfo")]
+    public TestScriptFileInfo? TestScriptFileInfo { get; set; }
+
+    [JsonPropertyName("additionalFileInfo")]
+    public List<AdditionalFileInfo>? AdditionalFileInfo { get; set; } = new();
+}