+++ src/Areas/LoadTesting/Models/LoadTest/PassFailCriteria.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace AzureMcp.Areas.LoadTesting.Models.LoadTest;
+
+public class PassFailCriteria
+{
+    [JsonPropertyName("passFailMetrics")]
+    public Dictionary<string, object>? PassFailMetrics { get; set; } = new();
+
+    [JsonPropertyName("passFailServerMetrics")]
+    public Dictionary<string, object>? PassFailServerMetrics { get; set; } = new();
+}