+++ src/Areas/LoadTesting/Models/LoadTestRun/TestRunCreateRequest.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+namespace AzureMcp.Areas.LoadTesting.Models.LoadTestRun;
+
+public class TestRunCreateRequest
+{
+    [JsonPropertyName("testRunRequest")]
+    public TestRunRequest TestRunRequest { get; set; } = new();
+}