+++ src/Program.cs
@@ -76,6 +76,7 @@ internal class Program
             new AzureMcp.Areas.Storage.StorageSetup(),
             new AzureMcp.Areas.BicepSchema.BicepSchemaSetup(),
             new AzureMcp.Areas.AzureTerraformBestPractices.AzureTerraformBestPracticesSetup(),
+            new AzureMcp.Areas.LoadTesting.LoadTestingSetup(),
         ];
     }